third_party/WebKit/LayoutTests/inspector/console/console-xpath.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/console-test.js"></script>
<script>

function test()
{
    InspectorTest.addConsoleViewSniffer(addMessageSniffer, true);

    InspectorTest.evaluateInConsole("$x('42')"); // number
    InspectorTest.evaluateInConsole("$x('name(/html)')"); // string
    InspectorTest.evaluateInConsole("$x('not(42)')"); // boolean
    InspectorTest.evaluateInConsole("$x('/html/body/p').length"); // node iterator
    InspectorTest.evaluateInConsole("$x('//a/@href')[0]"); // href, should not throw
    InspectorTest.evaluateInPage("console.log('complete')"); // node iterator

    function addMessageSniffer(uiMessage)
    {
        if (uiMessage.toString().indexOf("complete") !== -1) {
            InspectorTest.expandConsoleMessages(waitForObjects);
        }
    }

    function waitForObjects()
    {
        InspectorTest.waitForRemoteObjectsConsoleMessages(dumpConsoleMessages);
    }

    function dumpConsoleMessages()
    {
        InspectorTest.dumpConsoleMessages();
        InspectorTest.completeTest();
    }
}

</script>
</head>
<a href="http://chromium.org"></a>
<body onload="runTest()">
<p>
Tests $x for iterator and non-iterator types.
</p>

</body>
</html>
