third_party/WebKit/LayoutTests/inspector/console/console-xpath-expected.txt

CONSOLE MESSAGE: line 1: complete
Tests $x for iterator and non-iterator types.

$x('42')
$x('name(/html)')
$x('not(42)')
$x('/html/body/p').length
$x('//a/@href')[0]
42
"html"
false
1
href="http://chromium.org"
VM:1 complete
